somehow requesting user location permission works :D
